+++ CHANGELOG.md
@@ -1,6 +1,14 @@
+# 1.2.0
+- Added
+  - Export repos to check to YAML
+- Changed
+  - Plain text import changed for YAML
+  - Update dependencies
+
 # 1.1.1
 - Changed
   - GitHub Actions, rename, update versions, change the `set-output` for the new one
+
 # 1.1.0
 - Added
   - Pull request notifications in bulk import file